MPI tests should use (MAX_PROCS-1) peers and 1 server
These changes are necessary to support the ibverbs interface which cannot
have a client and server on the same node, so we should use 1 server and
(N-1) clients for testing to ensure compatibility of the test framework.

Use MPI_NUM_CLIENTS instead of MPIEXEC_MAX_NUMPROCS when setting up tests
to prevent ambiguity.